fix bug where numbers on the pathway mapper changes
